backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-dotenv
langgraph
langchain
langchain-core
langchain-google-genai
WEASYPRINT
stripe
